% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rePhotoFunctions.R
\name{AngleCalcs}
\alias{AngleCalcs}
\title{AngleCalcs}
\usage{
AngleCalcs(data1,out,CameraLens=90,UseOrder=TRUE)
}
\arguments{
\item{...}{Arguments are inherited from rePhotograph() and have the same characteristics, see note above about 'out'}
}
\value{
Returns 'out' the same output as rePhotograph()
}
\description{
A workhorse function that rePhotograph() uses to do the evaluate of each location as possible location for the camera.
}
\details{
This function is separated from the core rePhotograph function because it enables the user to verify the trueness of a prediction of search area if a putative camera location is found through ground searching. For instance, typically 'out' is a large data frame for evaluation, but replacing 'out' with out=dataframe(Lat=40,Long=50) TEST THIS would provide an evaluation of a specific location.
}
\examples{
see rePhotograph vignette
}
