% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rePhotoFunctions.R
\name{LineResid}
\alias{LineResid}
\title{LineResid}
\usage{
LineResid(Lat1,Lat2,m,c,Lat,Long)
}
\arguments{
\item{Lat1}{value; latitude of one POR}

\item{Lat2}{value; latitude of another POR. Importantly, the POR for Lat2 is assumed to be to the right of the POR for Lat1 (verified internally to rePhotograph)}

\item{m}{value; slope of the line connecting the two POR}

\item{c}{value; intercept of the line connecting the two POR}

\item{Lat}{value or vector; latitudes of one or many locations to be evaluated}

\item{Long}{value or vector; longitudes of one or many locations to be evaluated}
}
\value{
residuals a vector of values the length of Lat and Long inputs. Values of 0 signify that the Lat and Long inputs are on the opposite side of a the line connectin Lat1 and Lat2
}
\description{
A housekeeping function that AngleCalcs() uses to exclude locations as being on the wrong side of a line connecting two POR i.e. will return a value of 0 if a given latitude and longitude are on the far side of a line to the actual location. The function is capable of using vectors of latitudes and longitudes, or single values.
}
\details{
The function is based upon this logic: 1) a location exists where a photograph was taken, 2) two points of reference are ordered from left to right in a photograph and relative to that location, 3) there is a geographic line connecting the POR. Based upon these conditions, then the residuals of any arbitrary location relative to the line are diagnostic of whether the location is closer to the actual location than the line. This provides a means of excluding an arbitrary location as a possible location for the position of the camera i.e. if it is further than the line connecting the POR it can be excluded as the order of points would reverse.
}
\examples{
not used by user
}
