% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rePhotoFunctions.R
\name{outputKML}
\alias{outputKML}
\title{outputKML}
\usage{
outputKML(out,filename="yourfilename",Thres=0.05)
}
\arguments{
\item{out}{the output of rePhotograph()}

\item{filename}{the name of the output KML file}

\item{Thres}{the threshold for the contour output i.e. the default, Thres = 0.05, includes in the contour all locations that have a out$diffcomb < 0.05. This value can be adjusted based upon the regions shown in plot.search.area()}
}
\value{
A KML file that can be opened in Google Earth, CalTopo etc
}
\description{
A function for output of a KML file for use with GIS software. The KML file shows a contour encompassing a search area for the location of the camera in the original photograph.
}
\details{
This function requires functions from the sp, rgdal and maptools packages
}
\examples{
see rePhotograph vignette
}
