% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rePhotoFunctions.R
\name{plot.search.area}
\alias{plot.search.area}
\title{plot.search.area}
\usage{
plot.search.area(data1,out,extrapoints=data.frame(Lat=0,Long=0),title="Search areas are represented by lower values")
}
\arguments{
\item{data1}{the data input into rePhotograph providing the locations of the three or four points of reference. These are plotted as +'s in the contour plot}

\item{out}{the data output from rePhotograph providing the data frame of locations and results;}

\item{extrapoints}{a data.frame of any length that contains latitude and longitudes; allows plotting of other reference points that may be used as guides or possibly to mark a guess at the photograph location.}

\item{title}{a character vector providing a title for the plot e.g. a general location}
}
\value{
A filled.contour plot conditioned to show relevant aspects of the output of rePhotograph. The 'out' data frame is conditioned into a matrix and plotted as contours of locations that have similar 'diffcomb' values. These regions indicate the likely areas in which the original photograph was likely taken, lower values being more likely.
}
\description{
Used to plot the output of rePhotograph(), showing regions of similar differences i.e. regions where the original photograph was likely to be taken from. The main uses are to visualize the output of rePhotograph() and use these to change the default settings to narrow the result on one area e.g. by changing the grid or scaleToSearch or PtDensity. An important use is to evaluate a suitable threshold to determine the search area output in the outputKML function. For instance, in the examples above a small area of the total grid has a diffcomb value of < 0.05, this is the default threshold for outputKML i.e. only that area will be output as a KML search area for use with GIS software. In the last figure above, a very small region has a diffcomb<0.01, and this could be used to produce a very small search area as is shown in the example above.
}
\details{
This function is relatively simple and could be unpacked and reformulated to provide a user specific output.
}
\examples{
see rePhotograph vignette
}
